+++ app/index.tsx
@@ -3,11 +3,11 @@ import { useShallow } from "zustand/react/shallow"
 
 import { Navbar } from "~app/component/navbar"
 import { ProviderContextProvider } from "~app/context/provider"
-import { Info } from "~app/pages/info"
-import { Send } from "~app/pages/send"
+import { Info } from "~app/page/info"
+import { Send } from "~app/page/send"
+import { Path } from "~app/path"
 import { useStorage } from "~app/storage"
 import { useHashLocation } from "~app/util/location"
-import { PopupPath } from "~app/util/page"
 
 import "~style.css"
 
@@ -23,11 +23,11 @@ function App() {
       <Navbar />
       <Router hook={useHashLocation}>
         <Switch>
-          <Route path={PopupPath.root}>
-            <Redirect to={PopupPath.info} />
+          <Route path={Path.root}>
+            <Redirect to={Path.info} />
           </Route>
-          <Route path={PopupPath.info} component={Info}></Route>
-          <Route path={PopupPath.send} component={Send}></Route>
+          <Route path={Path.info} component={Info}></Route>
+          <Route path={Path.send} component={Send}></Route>
         </Switch>
       </Router>
     </ProviderContextProvider>
