+++ app/page/info.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState, type MouseEvent } from "react"
 import { Link } from "wouter"
 
 import { useProviderContext } from "~app/context/provider"
+import { Path } from "~app/path"
 import { useAccount } from "~app/storage"
-import { PopupPath } from "~app/util/page"
 import type { HexString } from "~typing"
 
 export function Info() {
@@ -21,9 +21,6 @@ export function Info() {
     const asyncFn = async () => {
       const explorer = getExplorerUrl((await provider.getNetwork()).name)
       setExplorerUrl(explorer)
-
-      // const [account] = await provider.listAccounts()
-      // setAddress(account.address)
 
       const balance = await provider.getBalance(account.address)
       setBalance(balance)
@@ -128,7 +125,7 @@ const AccountInternalTransactions: React.FC<{
 const SwitchToSendPage: React.FC = () => {
   return (
     <div className="flex-col justify-center items-center h-auto p-3 border-0 rounded-lg text-base">
-      <Link href={PopupPath.send}>Send ↗</Link>
+      <Link href={Path.send}>Send ↗</Link>
     </div>
   )
 }
