+++ app/page/send.tsx
@@ -3,8 +3,8 @@ import { useCallback, useState, type ChangeEvent } from "react"
 import { Link } from "wouter"
 
 import { useProviderContext } from "~app/context/provider"
+import { Path } from "~app/path"
 import { useAccount } from "~app/storage"
-import { PopupPath } from "~app/util/page"
 import type { BigNumberish, HexString } from "~typing"
 
 export function Send() {
@@ -85,7 +85,7 @@ export function Send() {
           className="flex-1">
           Send
         </button>
-        <Link href={PopupPath.info} className="flex-1">
+        <Link href={Path.info} className="flex-1">
           Cancel
         </Link>
       </div>
