+++ app/util/location.ts
@@ -1,4 +1,3 @@
-import { useEffect } from "react"
 import {
   navigate,
   useLocationProperty,
